requirements.txt
absl-py==2.0.0
aiohttp @ file:///croot/aiohttp_1694181079612/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs @ file:///croot/attrs_1695717823297/work
Babel==2.13.1
backcall==0.2.0 
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker @ file:///croot/blinker_1696539051175/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646038465422/work
comm==0.2.0
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.10
docopt==0.6.2
easydict==1.13
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.44.3
fqdn==1.5.1
frozenlist @ file:///croot/frozenlist_1670004507010/work
fsspec==2023.4.0
future==0.18.3
gast==0.3.3
google-auth @ file:///croot/google-auth_1694152708165/work
google-auth-oauthlib==1.0.0
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
graphviz==0.20.1
grpcio==1.59.2
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
idna @ file:///croot/idna_1666125576474/work
image==1.5.33
imageio==2.32.0
importlib-metadata @ file:///croot/importlib-metadata_1678997070253/work
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
jax==0.4.13
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.2
keras==2.13.1
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Mako==1.2.4
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
multidict @ file:///croot/multidict_1665674239670/work
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib @ file:///croot/oauthlib_1679489621486/work
opencv-python==4.9.0.80
opencv-python-headless==4.8.1.78
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
overrides==7.4.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs @ file:///croot/platformdirs_1692205439124/work
pooch @ file:///croot/pooch_1695850093751/work
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycuda==2022.2.2
pydot==1.4.2
Pygments==2.17.1
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytools==2023.1.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
shapely==2.0.3
six==1.15.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX==2.6.2.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.18.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==2.2.1
torch-tb-profiler==0.4.3
torchaudio==2.1.1+cu118
torchsummary==1.5.1
torchvision==0.17.1
torchviz==0.0.2
tornado==6.3.3
traitlets==5.13.0
trimesh==4.0.5
triton==2.2.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug @ file:///croot/werkzeug_1679489717957/work
widgetsnbextension==4.0.9
wrapt==1.12.1
yarg==0.1.9
yarl @ file:///opt/conda/conda-bld/yarl_1661437085904/work
zipp @ file:///croot/zipp_1672387121353/work
